mengembalikan float menjadi int

+++ insertion sort/insertion sort.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 using namespace std;
 
-float arr[20];            //Membuat array dengan panjang data 20
-float n;                 //Membuat Variable inputan n
+int arr[20];            //Membuat array dengan panjang data 20
+int n;                 //Membuat Variable inputan n
 
 void input() {          //Procedure Input
     while (true) {
